requirements.txt
firebase-admin
google-adk
yfinance
psutil
litellm
google-generativeai
python-dotenv
deprecated
